+++ app/src/main/java/com/example/tyler/calculatornew/SubtractInteger.java
@@ -1,18 +1,18 @@
 package com.example.tyler.calculatornew;
 
-public class SubtractInteger extends BinaryOperator<Integer>{
+public class SubtractInteger extends BinaryOperator<Double>{
 
     @Override
-    protected Integer calc(Integer a, Integer b){
+    protected Double calc(Double a, Double b){
         double total = a-b;
 
-        if (total > Integer.MAX_VALUE){
+        if (total > Double.MAX_VALUE){
             throw new ArithmeticException("Integer overflow");
         }
-        if (total < Integer.MIN_VALUE){
+        if (total < Double.MIN_VALUE){
             throw new ArithmeticException("Integer underflow");
         }
-        return (int) total;
+        return total;
     }
 
 }
